#20: Outlined all tests and their dependencies to a separate module
