+++ oo-memoized/memoized-chm/pom.xml
@@ -14,5 +14,9 @@
             <groupId>com.pragmaticobjects.oo.memoized</groupId>
             <artifactId>memoized-core</artifactId>
         </dependency>
+        <dependency>
+            <groupId>org.slf4j</groupId>
+            <artifactId>slf4j-api</artifactId>
+        </dependency>
     </dependencies>
 </project>
